"Refactored routes and server code, removed tryCatcher and swagger routes, updated pharmacy and users routes, and reorganized server.js"
