+++ config/passport-setup.js
@@ -1,0 +1,25 @@
+const GitHubStrategy = require("passport-github2").Strategy;
+
+async function githubAuth(passport) {
+    passport.use(
+    new GitHubStrategy(
+      {
+        clientID: process.env.GITHUB_CLIENT_ID,
+        clientSecret: process.env.GITHUB_CLIENT_SECRET,
+        callbackURL: process.env.CALLBACK_URL,
+      },
+      function (accessToken, refreshToken, profile, done) {
+        return done(null, profile);
+      }
+    )
+  );
+
+  passport.serializeUser((user, done) => {
+    done(null, user);
+  });
+  passport.deserializeUser((user, done) => {
+    done(null, user);
+  });
+}
+
+module.exports = { githubAuth };
