+++ routes/users.js
@@ -1,9 +1,9 @@
 const express = require("express");
 const router = express.Router();
-const userController = require('../controller/users');
+const userController = require("../controller/users");
 //const {tryCatcher} = require("../errors/tryCatch");
-const {saveUser} = require('../middleware/validate');
-const {isAuthenticated} = require('../middleware/authenticate');
+const { saveUser } = require("../middleware/validate");
+const { isAuthenticated } = require("../middleware/authenticate");
 
 /**
  * @route   GET /users
